+++ packages/subgraph/tests/token/governance-erc20.test.ts
@@ -9,6 +9,7 @@ import {
   ADDRESS_TWO,
   ONE_ETH,
   ADDRESS_THREE,
+  DAO_TOKEN_ADDRESS,
 } from '../constants';
 import {ExtendedTokenVotingMember} from '../helpers/extended-schema';
 import {createNewERC20TransferEvent, createTokenVotingMember} from './utils';
@@ -156,6 +157,9 @@ describe('Governance ERC20', () => {
 
       handleDelegateChanged(event);
 
+      // expected changes
+      member.delegatee = [memberAddress, pluginAddress].join('_');
+
       member.assertEntity();
       assert.entityCount('TokenVotingMember', 1);
     });
@@ -275,7 +279,7 @@ describe('Governance ERC20', () => {
       assert.entityCount('TokenVotingMember', 1);
     });
 
-    test('it should delete a member without voting power or balance', () => {
+    test('it should delete a member without voting power and balance and not delegating to another address', () => {
       let memberOneAddress = ADDRESS_ONE;
       let memberTwoAddress = ADDRESS_TWO;
       let pluginAddress = ADDRESS_SIX;
@@ -287,7 +291,7 @@ describe('Governance ERC20', () => {
         memberTwoAddress,
         pluginAddress
       );
-      /* member one has 100s token delegated to member two*/
+      /* member one has 100 token delegated to member two*/
       memberOne.balance = BigInt.fromString('100');
       memberOne.votingPower = BigInt.fromString('0');
       /* member two balance is 0 but has 100 voting power from the delegation of member one */
@@ -300,9 +304,15 @@ describe('Governance ERC20', () => {
 
       assert.entityCount('TokenVotingMember', 2);
 
-      // member one undelegates from member two
+      // member one un-delegates from member two
       let eventOne = memberOne.createEvent_DelegateVotesChanged('100');
       let eventTwo = memberTwo.createEvent_DelegateVotesChanged('0');
+
+      memberTwo.mockCall_delegatesCall(
+        DAO_TOKEN_ADDRESS,
+        ADDRESS_TWO,
+        ADDRESS_TWO
+      );
 
       handleDelegateVotesChanged(eventOne);
       handleDelegateVotesChanged(eventTwo);
@@ -311,9 +321,58 @@ describe('Governance ERC20', () => {
       // expected changes
       memberOne.votingPower = BigInt.fromString('100');
       memberOne.assertEntity();
-      // member two should be deleted because it has no balance or voting power
+      // member two should be deleted because it has no (balance and voting power) and not delegates to another address.
       assert.notInStore('TokenVotingMember', memberTwo.id);
       assert.entityCount('TokenVotingMember', 1);
+    });
+
+    test('it should not delete a member without voting power and balance, but delegating to another address', () => {
+      let memberOneAddress = ADDRESS_ONE;
+      let memberTwoAddress = ADDRESS_TWO;
+      let pluginAddress = ADDRESS_SIX;
+      let memberOne = new ExtendedTokenVotingMember().withDefaultValues(
+        memberOneAddress,
+        pluginAddress
+      );
+      let memberTwo = new ExtendedTokenVotingMember().withDefaultValues(
+        memberTwoAddress,
+        pluginAddress
+      );
+      /* member one has 100 token delegated to member two*/
+      memberOne.balance = BigInt.fromString('100');
+      memberOne.votingPower = BigInt.fromString('0');
+      /* member two balance is 0 but has 100 voting power from the delegation of member one */
+      memberTwo.balance = BigInt.fromString('0');
+      memberTwo.votingPower = BigInt.fromString('100');
+      /* member three has 100 tokens and none delegated */
+
+      memberOne.buildOrUpdate();
+      memberTwo.buildOrUpdate();
+
+      assert.entityCount('TokenVotingMember', 2);
+
+      // member one un-delegates from member two
+      let eventOne = memberOne.createEvent_DelegateVotesChanged('100');
+      let eventTwo = memberTwo.createEvent_DelegateVotesChanged('0');
+
+      memberTwo.mockCall_delegatesCall(
+        DAO_TOKEN_ADDRESS,
+        ADDRESS_TWO,
+        ADDRESS_ONE
+      );
+
+      handleDelegateVotesChanged(eventOne);
+      handleDelegateVotesChanged(eventTwo);
+
+      // assert
+      // expected changes
+      memberOne.votingPower = BigInt.fromString('100');
+      memberOne.assertEntity();
+
+      assert.fieldEquals('TokenVotingMember', memberOne.id, 'id', memberOne.id);
+      // memberTwo should not be deleted because it has no (balance and voting power), but it delegates to another address.
+      assert.fieldEquals('TokenVotingMember', memberTwo.id, 'id', memberTwo.id);
+      assert.entityCount('TokenVotingMember', 2);
     });
   });
 });